requirements.txt
fastapi
uvicorn
python-docx
psycopg2-binary
mammoth
beautifulsoup4
requests
